requirements.txt
discord.py
asyncio.py
random.py